+++ src/ecommerce/catalog/lifecycle/shippingmethods/EnvioExpressMock.java
@@ -1,7 +1,7 @@
 package ecommerce.catalog.lifecycle.shippingmethods;
 
 public class EnvioExpressMock {
-	public static float calcularCosto(float price) {
-		return 0f;
+	public static double calcularCosto(double price) {
+		return 0;
 	}
 }
